refactored clients from @/components to @/lib
