+++ src/components/Curators/RecordsAwaitingApproval.vue
@@ -160,7 +160,7 @@
             <td>
               {{ props.item.createdAt }},
               <a :href="'#/users/' + props.item.idCreator">
-                {{ props.item.creator }} 
+                {{ props.item.creator }}
               </a>
             </td>
           </tr>
@@ -267,7 +267,7 @@
 
 <script>
     import { mapActions, mapState }  from "vuex"
-    import RestClient from "@/components/Client/RESTClient.js"
+    import RestClient from "@/lib/Client/RESTClient.js"
     import Icon from "@/components/Icon"
 
     const restClient = new RestClient();
