+++ src/views/Records/Record.vue
@@ -133,7 +133,7 @@
 <script>
     import {mapActions, mapState, mapGetters, mapMutations} from 'vuex'
     import Client from '@/components/GraphClient/GraphClient.js'
-    import RestClient from "@/components/Client/RESTClient.js"
+    import RestClient from "@/lib/Client/RESTClient.js"
     import stringUtils from '@/utils/stringUtils';
     import GeneralInfo from "@/components/Records/Record/GeneralInfo";
     import Tools from '@/components/Records/Record/Tools';
