+++ src/views/Users/ConfirmAccount.vue
@@ -46,7 +46,7 @@
 </template>
 
 <script>
-    import RESTClient from "@/components/Client/RESTClient.js"
+    import RESTClient from "@/lib/Client/RESTClient.js"
     const client = new RESTClient();
 
     export default {
