+++ src/views/Users/ResendConfirmation.vue
@@ -69,7 +69,7 @@
 </template>
 
 <script>
-  import RESTClient from "@/components/Client/RESTClient.js"
+  import RESTClient from "@/lib/Client/RESTClient.js"
   let restClient = new RESTClient();
   import { isEmail, isRequired } from "@/utils/rules.js"
