+++ tests/unit/components/Editor/DataAccess/EditDataAccess.spec.js
@@ -3,7 +3,7 @@ import Vuex from "vuex"
 import Vuetify from "vuetify"
 import VueRouter from "vue-router"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
-import RestClient from "@/components/Client/RESTClient.js"
+import RestClient from "@/lib/Client/RESTClient.js"
 import DataAccess from "@/components/Editor/DataAccess/EditDataAccess.vue"
 import recordStore from "@/store/recordData.js"
 import editorStore from "@/store/editor.js"
