+++ tests/unit/components/Editor/DataAccess/EditSupportLinks.spec.js
@@ -2,7 +2,7 @@ import Vue from "vue"
 import { createLocalVue, shallowMount } from "@vue/test-utils"
 import Vuex from "vuex"
 import Vuetify from "vuetify"
-import ExternalClients from "@/components/Client/ExternalClients.js"
+import ExternalClients from "@/lib/Client/ExternalClients.js"
 import EditSupportLinks from "@/components/Editor/DataAccess/EditSupportLinks.vue"
 import recordStore from "@/store/recordData.js"
 import editorStore from "@/store/editor.js"
