+++ tests/unit/components/Editor/EditPublications.spec.js
@@ -7,8 +7,8 @@ import recordStore from "@/store/recordData.js";
 import editorStore from "@/store/editor.js";
 import userStore from "@/store/users.js";
 import GraphClient from "@/components/GraphClient/GraphClient.js";
-import RestClient from "@/components/Client/RESTClient.js"
-import ExternalClient from "@/components/Client/ExternalClients.js"
+import RestClient from "@/lib/Client/RESTClient.js"
+import ExternalClient from "@/lib/Client/ExternalClients.js"
 const sinon = require("sinon");
 
 const localVue = createLocalVue();
