+++ tests/unit/components/Editor/EditRelations.spec.js
@@ -6,7 +6,7 @@ import editRelations from "@/components/Editor/EditRelationships.vue"
 import recordStore from "@/store/recordData.js";
 import editorStore from "@/store/editor.js";
 import userStore from "@/store/users.js";
-import RestClient from "@/components/Client/RESTClient.js"
+import RestClient from "@/lib/Client/RESTClient.js"
 import GraphClient from "@/components/GraphClient/GraphClient.js";
 import relationTypes from "@/../tests/fixtures/relationTypes.json"
 const sinon = require("sinon");
