+++ tests/unit/components/Editor/GeneralInformation/GeneralInformation.spec.js
@@ -8,7 +8,7 @@ import recordStore from "@/store/recordData.js"
 import editorStore from "@/store/editor.js"
 import userStore from "@/store/users.js"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
-import RestClient from "@/components/Client/RESTClient.js"
+import RestClient from "@/lib/Client/RESTClient.js"
 import countriesQuery from "@/components/GraphClient/queries/getCountries.json"
 import typesQuery from "@/components/GraphClient/queries/getRecordsTypes.json"
 import tagsQuery from "@/components/GraphClient/queries/geTags.json"
