+++ tests/unit/components/Editor/Organisations/Organisations.spec.js
@@ -3,7 +3,7 @@ import Vuex from "vuex"
 import Vuetify from "vuetify"
 import VueRouter from "vue-router"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
-import RestClient from "@/components/Client/RESTClient.js"
+import RestClient from "@/lib/Client/RESTClient.js"
 import getOrganisationsTypesQuery from "@/components/GraphClient/queries/Organisations/getOrganisationTypes.json"
 import getOrganisationsQuery from "@/components/GraphClient/queries/Organisations/getOrganisations.json"
 import getGrantsQuery from "@/components/GraphClient/queries/Organisations/getGrants.json"
