+++ tests/unit/store/users.spec.js
@@ -1,5 +1,5 @@
 import { mutations, actions } from "@/store/users.js"
-import Client from "@/components/Client/RESTClient.js"
+import Client from "@/lib/Client/RESTClient.js"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
 import { initUserDataState } from "@/store/utils.js"
 import Vue from "vue"
