+++ tests/unit/views/CreateRecord/Editor.spec.js
@@ -5,7 +5,7 @@ import recordStore from "@/store/recordData.js";
 import userStore from "@/store/users.js";
 import editorStore from "@/store/editor.js"
 import GraphClient from "@/components/GraphClient/GraphClient.js";
-import RESTClient from "@/components/Client/RESTClient.js"
+import RESTClient from "@/lib/Client/RESTClient.js"
 import metaTemplate from "../../../fixtures/metaTemplate.json"
 import VueRouter from "vue-router";
 const sinon = require("sinon");
