+++ tests/unit/views/Users/Login/Login.spec.js
@@ -3,7 +3,7 @@ import VueRouter from "vue-router"
 import Vuex from "vuex"
 import sinon from "sinon"
 import Login from "@/views/Users/Login/Login.vue"
-import Client from "@/components/Client/RESTClient.js"
+import Client from "@/lib/Client/RESTClient.js"
 import usersStore from "@/store/users.js";
 
 const localVue = createLocalVue();
