+++ tests/unit/views/Users/PublicProfile.spec.js
@@ -3,8 +3,8 @@ import VueRouter from "vue-router"
 import Vuex from "vuex"
 import sinon from "sinon"
 import PublicProfile from "@/views/Users/PublicProfile"
-import Client from "@/components/Client/RESTClient.js"
-import ExternalClient from "@/components/Client/ExternalClients.js"
+import Client from "@/lib/Client/RESTClient.js"
+import ExternalClient from "@/lib/Client/ExternalClients.js"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
 import usersStore from "@/store/users";
 import editorStore from "@/store/editor";
