+++ tests/unit/views/Users/ResendConfirmation.spec.js
@@ -1,7 +1,7 @@
 import { shallowMount } from "@vue/test-utils";
 import sinon from "sinon";
 import ResendConfirmation from "@/views/Users/ResendConfirmation";
-import Client from "@/components/Client/RESTClient.js"
+import Client from "@/lib/Client/RESTClient.js"
 
 describe("ResendConfirmation.vue", function(){
     let wrapper;
